Use RET rather than <return> for terminal compatibility
When using Emacs 26.3 and Emacs 26.1 in a terminal, pressing the Enter
key when navigating the history just causes a newline to appear in the
minibuffer splitting the history display.

It appears that "<return>" is only generated by graphical Emacs. All the
other packages I looked at bound "RET" instead, which also works in
graphical Emacs so it appears to be better to use that instead.